fixup: on last PR i forgot to update the codegen
n.b. regenerating the domain classes with this has a dependency
resolution issue, it picks the old codegen. Not sure why, but since we
don't really need the full cpg domain classes committed here I'll take
that as a trigger to setup a simplified sample schema here which is
independent of cpg...