DiffGraph: allow simultaneous edge and node removal
Prior to this, if a user tried to delete a node and an adjacent edge at
the same time, we would run into this:
```
scala.MatchError: flatgraph.DiffGraphBuilder$RemoveEdge@7bcbf3a8 (of class flatgraph.DiffGraphBuilder$RemoveEdge)
	at flatgraph.DiffGraphApplier.splitUpdate$$anonfun$2(DiffGraphApplier.scala:228) ~[io.joern.flatgraph-core_3-0.1.25.jar:0.1.25]
```

Which was because the condition was on the `case RemoveEdge` filtered
out that case and there's no fallback `case _` or similar.
This makes it explicit and adds a test.